Removed clutter, put endpoint handlers into seperate packages
