+++ internal/team/TeamHandler.go
@@ -1,13 +1,12 @@
-package handler
+package team
 
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/NTNU-sondrbaa-2019/CLOUD-PROJECT/internal/view"
 	"log"
 	"net/http"
-	"strconv"
 	"strings"
-	"time"
 )
 
 type Team struct {
@@ -18,10 +17,6 @@ type Team struct {
 
 type User struct {
 	Username string
-}
-
-type Page struct{
-
 }
 
 func TeamHandler(w http.ResponseWriter, r *http.Request, title string) {
@@ -37,11 +32,11 @@ func TeamHandler(w http.ResponseWriter, r *http.Request, title string) {
 			case "users":
 				UsersTeamHandler(w,r,title)
 			case "results":
-				ErrorPage(w, "Not implemented", http.StatusNotImplemented)
+				view.ErrorPage(w, "Not implemented", http.StatusNotImplemented)
 			case "seasons":
-				ErrorPage(w, "Not implemented", http.StatusNotImplemented)
+				view.ErrorPage(w, "Not implemented", http.StatusNotImplemented)
 			default:
-				ErrorPage(w, "Not Found", http.StatusNotFound)
+				view.ErrorPage(w, "Not Found\n\ndsafsa", http.StatusNotFound)
 			}
 		} else {
 			// Search for team name urlPart[4]
@@ -66,16 +61,5 @@ func TeamHandler(w http.ResponseWriter, r *http.Request, title string) {
 }
 
 func UsersTeamHandler(w http.ResponseWriter, r *http.Request, title string) {
-	ErrorPage(w, "Not implemented", http.StatusNotImplemented)
-}
-
-func ErrorPage(w http.ResponseWriter,errorMsg string, code int) {
-	type err struct {
-		ErrorMsg string
-		ErrorCode int
-		CurrentYear string
-	}
-	error := &err{ErrorMsg: errorMsg, ErrorCode: code, CurrentYear: strconv.Itoa(time.Now().Year())}
-
-	RenderIndex(w, "error", error)
+	view.ErrorPage(w, "Not implemented", http.StatusNotImplemented)
 }
