+++ internal/user/UserHandler.go
@@ -1,4 +1,4 @@
-package handler
+package user
 
 import "net/http"
