+++ internal/view/config.go
@@ -2,4 +2,8 @@ package view
 
 import "html/template"
 
-var Templates = template.Must(template.ParseFiles("web/static/login.html","web/static/ucp.html","web/static/error.html"))
+var Templates = template.Must(template.ParseFiles(
+	"web/static/login.html",
+	"web/static/ucp.html",
+	"web/static/error.html",
+	))
